Move TupleSlim to its own file
